Implement rudimentary DictionaryArray::Validate
